Move AuditLog to org.oskari.log
