+++ control-admin/src/main/java/fi/nls/oskari/control/admin/CacheHandler.java
@@ -11,7 +11,7 @@ import fi.nls.oskari.log.LogFactory;
 import fi.nls.oskari.log.Logger;
 import fi.nls.oskari.util.JSONHelper;
 import fi.nls.oskari.util.ResponseHelper;
-import fi.nls.oskari.utils.AuditLog;
+import org.oskari.log.AuditLog;
 import org.json.JSONArray;
 import org.json.JSONObject;
