+++ control-base/src/main/java/fi/nls/oskari/control/data/AcceptPublishedTermsOfUseHandler.java
@@ -2,13 +2,12 @@ package fi.nls.oskari.control.data;
 
 import fi.nls.oskari.annotation.OskariActionRoute;
 import fi.nls.oskari.control.ActionException;
-import fi.nls.oskari.control.ActionHandler;
 import fi.nls.oskari.control.ActionParameters;
 import fi.nls.oskari.control.RestActionHandler;
 import fi.nls.oskari.map.publish.service.PublishTermsOfUseService;
 import fi.nls.oskari.map.publish.service.PublishTermsOfUseServiceMybatisImpl;
 import fi.nls.oskari.util.ResponseHelper;
-import fi.nls.oskari.utils.AuditLog;
+import org.oskari.log.AuditLog;
 
 @OskariActionRoute("AcceptPublishedTermsOfUse")
 public class AcceptPublishedTermsOfUseHandler  extends RestActionHandler {
