+++ control-base/src/main/java/fi/nls/oskari/control/view/SaveViewHandler.java
@@ -13,7 +13,7 @@ import fi.nls.oskari.map.view.ViewService;
 import fi.nls.oskari.map.view.AppSetupServiceMybatisImpl;
 import fi.nls.oskari.util.JSONHelper;
 import fi.nls.oskari.util.ResponseHelper;
-import fi.nls.oskari.utils.AuditLog;
+import org.oskari.log.AuditLog;
 import org.apache.commons.lang.StringEscapeUtils;
 import org.json.JSONObject;
