+++ control-myplaces/src/main/java/fi/nls/oskari/control/myplaces/handler/MyPlacesFeaturesHandler.java
@@ -5,7 +5,7 @@ import java.io.InputStream;
 import java.util.Arrays;
 import java.util.List;
 
-import fi.nls.oskari.utils.AuditLog;
+import org.oskari.log.AuditLog;
 import org.json.JSONException;
 import org.json.JSONObject;
