+++ control-statistics/src/main/java/fi/nls/oskari/control/statistics/user/AddIndicatorDataHandler.java
@@ -16,7 +16,7 @@ import fi.nls.oskari.service.OskariComponentManager;
 import fi.nls.oskari.util.JSONHelper;
 import fi.nls.oskari.util.PropertyUtil;
 import fi.nls.oskari.util.ResponseHelper;
-import fi.nls.oskari.utils.AuditLog;
+import org.oskari.log.AuditLog;
 import org.json.JSONObject;
 import org.oskari.statistics.user.StatisticalIndicatorService;
