+++ control-userlayer/src/main/java/org/oskari/control/userlayer/DeleteUserLayerHandler.java
@@ -1,13 +1,12 @@
 package org.oskari.control.userlayer;
 
 import fi.nls.oskari.control.RestActionHandler;
-import fi.nls.oskari.utils.AuditLog;
+import org.oskari.log.AuditLog;
 import org.json.JSONObject;
 import org.oskari.map.userlayer.service.UserLayerDbService;
 
 import fi.nls.oskari.annotation.OskariActionRoute;
 import fi.nls.oskari.control.ActionException;
-import fi.nls.oskari.control.ActionHandler;
 import fi.nls.oskari.control.ActionParameters;
 import fi.nls.oskari.domain.map.userlayer.UserLayer;
 import fi.nls.oskari.service.OskariComponentManager;
