+++ servlet-map/src/main/java/fi/nls/oskari/spring/security/OskariLoginFailureHandler.java
@@ -3,7 +3,7 @@ package fi.nls.oskari.spring.security;
 import fi.nls.oskari.control.ActionParameters;
 import fi.nls.oskari.log.LogFactory;
 import fi.nls.oskari.log.Logger;
-import fi.nls.oskari.utils.AuditLog;
+import org.oskari.log.AuditLog;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
